panel: caimito: avoid deadlock in drm_atomic_get_crtc_state
This atomic_set_property() is not designed to be a complete DRM call.

There is no need to provide full locking for it, as it only alters the locally
managed state object and a global matrix, and the state will never be committed
or mixed with other states. There is literally nothing that requires locking.

Furthermore, crtc will be locked by callers upon exiting LP mode, locking here
(via drm_atomic_get_crtc_state()) will lead to a deadlock.

This commit replaces drm_atomic_get_crtc_state() with a local variable to drop
the lock. Beware that its initialization is adequate only for the
linear_matrix_override call, do not use it for anything else.

Co-Authored-By: Hamster Tian <6023221+updateing@users.noreply.github.com>