panel: caimito: km4: specify features when sending dimming commands
In hk3_set_panel_feat(), the latest features are in a local variable, not
spanel->feat. This causes a mismatch between stock driver behavior and this
mod.

Add a feat argument for hk3_set_{override,default}_dimming() so that caller can
specify features, rather than relying on the implication about spanel->feat.

Co-Authored-By: Hamster Tian <6023221+updateing@users.noreply.github.com>